requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.6
asyncpg==0.23.0
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.0
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
limits==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
prometheus-client==0.11.0
prometheus-fastapi-instrumentator==5.7.1
pydantic==1.8.2
pylint==2.9.6
python-dotenv==0.19.0
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
six==1.16.0
slowapi==0.1.4
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
tomli==1.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
wrapt==1.12.1
